deprecating unused graphql
