--- frontend/agents/graphql/AgentByIdQuery.js
@@ -1,19 +1,0 @@
-import { graphql } from "react-relay/hooks";
-
-export const AgentByIdQuery = graphql`
-  query AgentByIdQuery($id: UUID!) {
-    agent(id: $id) {
-      id
-      name
-      alias
-      model
-      purpose
-      config
-      chain {
-        id
-        name
-        description
-      }
-    }
-  }
-`;
