--- frontend/agents/graphql/AgentsQuery.js
@@ -1,20 +1,0 @@
-import React from "react";
-import { graphql } from "relay-runtime";
-
-export const AgentsQuery = graphql`
-  query AgentsQuery {
-    agents {
-      id
-      name
-      alias
-      model
-      purpose
-      config
-      chain {
-        id
-        name
-        description
-      }
-    }
-  }
-`;
