--- frontend/chains/graphql/AddChainEdgeMutation.js
@@ -1,19 +1,0 @@
-import { graphql } from "react-relay";
-
-export const AddChainEdgeMutation = graphql`
-  mutation AddChainEdgeMutation($data: ChainEdgeInput!) {
-    addChainEdge(data: $data) {
-      edge {
-        id
-        key
-        inputMap
-        source {
-          id
-        }
-        target {
-          id
-        }
-      }
-    }
-  }
-`;
