--- frontend/chains/graphql/AddChainNodeMutation.js
@@ -1,22 +1,0 @@
-import { graphql } from "react-relay";
-
-export const AddChainNodeMutation = graphql`
-  mutation AddChainNodeMutation($data: ChainNodeInput!) {
-    addChainNode(data: $data) {
-      node {
-        id
-        classPath
-        config
-        name
-        description
-        position {
-          x
-          y
-        }
-        chain {
-          id
-        }
-      }
-    }
-  }
-`;
