--- frontend/chains/graphql/ChainByIdQuery.js
@@ -1,12 +1,0 @@
-import { graphql } from "react-relay/hooks";
-
-export const ChainByIdQuery = graphql`
-  query ChainByIdQuery($id: UUID!) {
-    chain(id: $id) {
-      id
-      name
-      description
-      createdAt
-    }
-  }
-`;
