--- frontend/chains/graphql/ChainsQuery.js
@@ -1,12 +1,0 @@
-import React from "react";
-import { graphql } from "relay-runtime";
-
-export const ChainsQuery = graphql`
-  query ChainsQuery {
-    chains {
-      id
-      name
-      description
-    }
-  }
-`;
