--- frontend/chains/graphql/DeleteChainEdgeMutation.js
@@ -1,11 +1,0 @@
-import { graphql } from "react-relay";
-
-export const DeleteChainEdgeMutation = graphql`
-  mutation DeleteChainEdgeMutation($id: UUID!) {
-    deleteChainEdge(id: $id) {
-      edge {
-        id
-      }
-    }
-  }
-`;
