--- frontend/chat/graphql/AddAgentMutation.js
@@ -1,14 +1,0 @@
-import { graphql } from "react-relay";
-
-export const AddAgentMutation = graphql`
-  mutation AddAgentMutation($chatId: UUID!, $agentId: UUID!) {
-    addAgent(chatId: $chatId, agentId: $agentId) {
-      chat {
-        id
-        agents {
-          id
-        }
-      }
-    }
-  }
-`;
