--- frontend/chat/graphql/RemoveAgentMutation.js
@@ -1,16 +1,0 @@
-import { graphql } from "relay-runtime";
-
-const RemoveAgentMutation = graphql`
-  mutation RemoveAgentMutation($chatId: UUID!, $agentId: UUID!) {
-    removeAgent(chatId: $chatId, agentId: $agentId) {
-      chat {
-        id
-        agents {
-          id
-        }
-      }
-    }
-  }
-`;
-
-export default RemoveAgentMutation;
