--- frontend/chat/graphql/useSendInput.js
@@ -1,35 +1,0 @@
-import { useCallback, useState } from "react";
-import axios from "axios";
-
-export const useSendInput = (chat_id) => {
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState(null);
-  const url = `/api/chats/${chat_id}/messages`;
-
-  const sendInput = useCallback(
-    async (text) => {
-      const data = {
-        chat_id: chat_id,
-        text: text,
-      };
-
-      setIsLoading(true);
-      try {
-        const response = await axios.post(url, data);
-        setIsLoading(false);
-        return response.data;
-      } catch (err) {
-        setIsLoading(false);
-        setError(err);
-        throw err;
-      }
-    },
-    [chat_id]
-  );
-
-  return {
-    sendInput,
-    isLoading,
-    error,
-  };
-};
